chore(erigon): remove dead code merged in #92
PR #92 merged a few dead/disposable items alongside the (live, load-bearing)
pure-Go snapshot writer. This trims them without touching the writer or
weakening cross-client correctness coverage.

- Relocate commitment.ComputeGenesisRootFromAccounts to a test-only helper
  (commitment/helpers_test.go). It is a test-only backward-compat wrapper
  with no production callers; production uses the streaming
  ComputeGenesisRoot. Removes a dead export from the production API.
- Delete client/erigon/parallel_equivalence_test.go. TestParallelEquivalence
  was excluded from the CI -run filter and the Makefile suite omits the
  cgo_erigon_commitment tag, so it was compiled but never executed.
- Remove the now-orphaned setErigonWorkers test hook from snapshot_cgo.go
  (its only caller was parallel_equivalence_test.go).
- Delete internal/erigon/recsplit/large_test.go (gated behind the
  recsplit_large tag that no CI job or Makefile target sets).
- Remove the pre-existing unused keyLess helper (and its now-orphaned bytes
  import) from recsplit/spill.go.
- Trim recsplit testdata: drop spike_10000.json (40k lines / 1.2 MB). The
  spike_100/spike_1000 pair still covers the leaf-bijection and
  recursive-split encoder paths; at-scale coverage comes from the
  real-daemon TestE2ESuite.

Kept (load-bearing, on the production --client=erigon path): the
internal/erigon/* snapshot writer, account/serialise_v3, and the
btindex/seg/existence golden parity gates. The writer is called
unconditionally (run_cgo.go:156) and the golden tests gate byte-format
parity vs Erigon in the default pure-Go CI job.

Verified: go build ./internal/erigon/..., go test ./internal/erigon/recsplit/
green; gofmt clean. The cgo-gated edits (commitment relocation, snapshot_cgo,
parallel-test deletion) are verified by inspection + gofmt; the cgo CI leg is
authoritative (Docker not runnable locally).
