+++ internal/erigon/commitment/commitment.go
@@ -243,52 +243,6 @@ func ComputeGenesisRoot(commitmentInputStore *streamsort.Store) (Result, error) 
 	return Result{Root: root, BranchNodes: mergedBranches, HPHState: hphState}, nil
 }
 
-// ComputeGenesisRootFromAccounts is a backward-compat wrapper for
-// small in-memory inputs (tests + the H4 invariance proof). Materialises
-// the slice into a temp streamsort + calls the streaming
-// ComputeGenesisRoot. Not for production at bench scale.
-func ComputeGenesisRootFromAccounts(accounts []Account) (Result, error) {
-	store, err := streamsort.New("")
-	if err != nil {
-		return Result{}, fmt.Errorf("ComputeGenesisRootFromAccounts: streamsort.New: %w", err)
-	}
-	defer store.Close()
-
-	for _, a := range accounts {
-		// Account entry keyed by 20-byte address.
-		var balance *uint256.Int
-		if a.Balance != nil {
-			balance = a.Balance
-		}
-		acctBytes := EncodeAccountUpdate(a.Nonce, balance, a.Code)
-		if err := store.Put(a.Address[:], acctBytes); err != nil {
-			return Result{}, fmt.Errorf("ComputeGenesisRootFromAccounts: put account %s: %w", a.Address.Hex(), err)
-		}
-		// Storage entries keyed by addr||slot. Skip all-zero values.
-		for slot, val := range a.Storage {
-			trimmed := trimLeadingZeros(val[:])
-			if len(trimmed) == 0 {
-				continue
-			}
-			composite := make([]byte, 0, 52)
-			composite = append(composite, a.Address[:]...)
-			composite = append(composite, slot[:]...)
-			storBytes := EncodeStorageUpdate(val[:])
-			if err := store.Put(composite, storBytes); err != nil {
-				return Result{}, fmt.Errorf("ComputeGenesisRootFromAccounts: put storage %s/%s: %w", a.Address.Hex(), slot.Hex(), err)
-			}
-		}
-	}
-	// ComputeGenesisRoot requires its input streamsort to be Finalized
-	// — Iterate and Get on the store both gate on the Finalize state
-	// transition. The wrapper Puts everything here, so we Finalize
-	// before delegating.
-	if err := store.Finalize(); err != nil {
-		return Result{}, fmt.Errorf("ComputeGenesisRootFromAccounts: Finalize: %w", err)
-	}
-	return ComputeGenesisRoot(store)
-}
-
 // subtreeCtx implements erigoncommitment.PatriciaContext over a
 // streamsort-backed commitmentInputStore (random-access via Pebble,
 // thread-safe read path) plus a PER-WORKER branches map.
